fix spiral heater

+++ gdsfactory/components/bend_euler.py
@@ -75,8 +75,8 @@ def bend_euler(
     c.add_ports(ref.ports)
     c.info["length"] = np.round(p.length(), 3)
     c.info["dy"] = np.round(abs(float(p.points[0][0] - p.points[-1][0])), 3)
-    c.info["radius_min"] = np.round(p.info["Rmin"], 3)
-    c.info["radius"] = radius
+    c.info["radius_min"] = float(np.round(p.info["Rmin"], 3))
+    c.info["radius"] = float(radius)
 
     if direction == "cw":
         ref.mirror(p1=[0, 0], p2=[1, 0])
